Added compression capabilities.
Moved the compression functions from the libsnapwebsites environment.
Added support for bz2 compression.
Wrote 100% coverage for the compression functions and fixed a few bugs.
Started to tweaked the exceptions to make them common instead of specific to each section of the library.
Fixed error message in is_token() since tokens are not just in cookies.
Avoid "." as path segments. This makes it simpler to detect invalid ".." segments and it's better canonicalization.
Made the hashbang path a separate path (I'm thinking it should just go away now).
Fixed a bug in the implementation of the query string when a name is missing.
Fixed a bug where the ':' between the domain name and port was not being added on a get().
Added "is-unix" as one of the part we can query.
Moved the "scheme" to the 's' case in the part switch.
Fixed the port when read from the /etc/service file, it comes in network order.